pyproject.toml
[tool.poetry]
name = "ux_autobot"
version = "0.1.0"
description = ""
authors = ["Kenny B"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.0.0"
browsermob-proxy = "^0.8.0"
pyyaml = "^5.4.1"
webdriver-manager = "^3.5.1"
psutil = "^5.8.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
pylint = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"